Moved endpoints into controller files (out of folder)
